Fix(English): fix api delete duplicate

+++ src/study-program/quarter-report/src/Repositories/Eloquent/QuarterReportRepositoryEloquent.php
@@ -118,7 +118,7 @@ class QuarterReportRepositoryEloquent extends BaseRepository implements QuarterR
         if (!empty($attributes['limit'])) {
             $student = $this->studentRepositoryEloquent->paginate($attributes['limit']);
         } else {
-            $student = $this->studentRepositoryEloquent->get();
+            $student = $this->studentRepositoryEloquent->paginate(50);
         }
 
         return $student;
@@ -132,6 +132,7 @@ class QuarterReportRepositoryEloquent extends BaseRepository implements QuarterR
                 $attributes['reportTime'] = date('Y-m-d H:i:s');
                 $attributes['type'] = QuarterReport::TYPE['DUPLICATE'];
 
+                $quarterReportId = '';
                 for ($i = 1; $i <= 2; $i++) {
                     switch ($i) {
                         case 1:
@@ -139,9 +140,11 @@ class QuarterReportRepositoryEloquent extends BaseRepository implements QuarterR
                             break;
                         case 2:
                             $attributes['status'] = QuarterReport::STATUS['NOT_YET_CONFIRM'];
+                            $attributes['quarterReportId'] = $quarterReportId;
                             break;
                     }
                     $result = $this->model()::create($attributes);
+                    $quarterReportId = $result->Id;
                 }
             } else {
                 $result = $this->model()::create($attributes);
@@ -254,24 +257,10 @@ class QuarterReportRepositoryEloquent extends BaseRepository implements QuarterR
 
     public function deleteAll($id)
     {
-        $data = $this->model()::find($id);
-        DB::beginTransaction();
-        try {
-            foreach ($data->quarterReportDetail as $key => $value) {
-                foreach ($value->quarterReportDetailSubject as $key => $valueDetail) {
-                    $valueDetail->quarterReportDetailSubjectChildren()->delete();
-                }
-                $value->quarterReportDetailSubject()->delete();
-            }
-            $data->quarterReportDetail()->delete();
-            $data->delete();
-            DB::commit();
-        } catch (\Throwable $th) {
-            DB::rollBack();
-            throw new HttpException(500, $th->getMessage());
-        }
+        $data = $this->model()::findOrFail($id);
+        $data->forceDelete();
 
-        return parent::all();
+        return parent::parserResult($this->model->orderBy('LastModificationTime', 'desc')->first());
     }
 
     public function updateStatusQuarterReport(array $attributes)
@@ -396,11 +385,20 @@ class QuarterReportRepositoryEloquent extends BaseRepository implements QuarterR
 
     public function deleteQuarterReport($id)
     {
-        $quarterReport = $this->model()::withTrashed->find($id);
-        $duplicate = $this->model()::where('Type', QuarterReport::TYPE['DUPLICATE']);
-        $duplicate->forceDelete();
-        $quarterReport->forceDelete();
+        $quarterReport = $this->model()::findOrFail($id);
+        DB::beginTransaction();
+        try {
+            $duplicate = $this->model()::where('Id', $quarterReport->QuarterReportId)->first();
 
-        return parent::all();
+            if (!is_null($duplicate)) {
+                $duplicate->forceDelete();
+            }
+            $quarterReport->forceDelete();
+            DB::commit();
+        } catch (\Throwable $th) {
+            DB::rollBack();
+        }
+
+        return parent::parserResult($this->model->orderBy('LastModificationTime', 'desc')->first());
     }
 }
